Reworked projects home page and new project pages
Reworked the home projects page so that its better and looks decent now. Still a little bit more work to do on that page but much better than it was. Also updated the buttons to use docusaurus's Links buttons (these are responsive which fixes issues on phones and other small screens). Also removed the light mode option and the theme toggle on the site as its my site and you will use dark mode. Lastly add some new project pages