Style improvements plus business card
Got buttons and project cards to style that I like. Fixed an issue on the about me page where the quote would cause the text to jump if it was long enough. Lastly recreated the business card as a page from my last site. Note completely happy with it currently so a little more work to do there